refactor(libs): rename peripheral libs with adc_/pwm_ prefix and add dev tooling
Rename ADC sensor libraries (ldr → adc_ldr, ntc → adc_ntc) and PWM
libraries (led_rgb → pwm_led_rgb, servo_pwm → pwm_servo_rc) to follow
consistent module_function naming convention. Add new header files for
each renamed module.

- Librerias/adc_ldr.c/h: renamed from ldr.c/h, updated Doxygen [NEW headers]
- Librerias/adc_ntc.c/h: renamed from ntc.c/h, updated Doxygen [NEW headers]
- Librerias/pwm_led_rgb.c/h: renamed from led_rgb.c/h [NEW header]
- Librerias/pwm_servo_rc.c/h: renamed from servo_pwm.c/h [NEW header]
- Librerias/main_adc_ldr.c, main_adc_ntc.c, main_pwm_led_rgb.c,
  main_pwm_servo_rc.c: renamed to match new library names
- Librerias/glcd.c/h: fuente parameter uint32_t → uint8_t
- Librerias/sdram.c: formatting cleanup
- Librerias/clases_*.h, leds.h, uart_lpc40xx.h: include and style updates
- Librerias/main_base.c, main_uart.c, main_test_funciones_macro.c: style fixes
- .clang-tidy, .clangd, .editorconfig, c_cpp_properties.json: [NEW] dev tooling
- .clang-format, estiloC.md, docDoxygen.md: style guide updates
- .github/workflows/mkdocs-ci.yml: CI workflow updates
- 02_Base.uvprojx, *.uvoptx: Keil project references updated

IMPROVEMENTS:
- Consistent peripheral prefix naming (adc_, pwm_) improves discoverability
- Separate .h files for each module follow project conventions
- Dev tooling (clang-tidy, clangd, editorconfig) enables static analysis
- Narrower fuente type (uint8_t) matches actual value range

+++ Librerias/glcd.c
@@ -549,7 +549,7 @@ void glcd_activar_desplazamiento(bool_t activar) {
   texto_actual.desplazamiento_activado = activar;
 }
 
-void glcd_seleccionar_fuente(uint32_t fuente) {
+void glcd_seleccionar_fuente(uint8_t fuente) {
 
   ASSERT(fuente < NUMERO_FUENTES, "Fuente incorrecta.");
 
@@ -568,7 +568,7 @@ int32_t glcd_printf(const char *format, ...) {
   return retval;
 }
 
-int32_t glcd_xprintf(uint16_t x, uint16_t y, uint16_t color, uint16_t color_fondo, uint32_t fuente,
+int32_t glcd_xprintf(uint16_t x, uint16_t y, uint16_t color, uint16_t color_fondo, uint8_t fuente,
                      const char *format, ...) {
 
   uint16_t pos_x_anterior, pos_y_anterior, fuente_anterior, color_anterior, color_fondo_anterior;
@@ -601,7 +601,7 @@ int32_t glcd_xprintf(uint16_t x, uint16_t y, uint16_t color, uint16_t color_fond
 }
 
 void glcd_caracter(char c, uint16_t x, uint16_t y, uint16_t color, uint16_t color_fondo,
-                   uint32_t fuente) {
+                   uint8_t fuente) {
 
   const fuente_t *fuente_ptr = tabla_fuentes[fuente];
   uint16_t pixeles_ancho, pixeles_alto, bytes_por_fila, salto_en_fb, i, j;
@@ -650,7 +650,7 @@ void glcd_caracter(char c, uint16_t x, uint16_t y, uint16_t color, uint16_t colo
   }
 }
 
-void glcd_texto(uint16_t x, uint16_t y, uint16_t color, uint16_t color_fondo, uint32_t fuente,
+void glcd_texto(uint16_t x, uint16_t y, uint16_t color, uint16_t color_fondo, uint8_t fuente,
                 const char *str) {
 
   const fuente_t *fuente_ptr = tabla_fuentes[fuente];
